rag_service/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
mysql-connector-python==8.2.0
openai==1.7.0

# -- LangChain Packages --
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
faiss-cpu==1.7.4

# Additional utilities
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
